+++ node/pkg/watchers/ibc/watcher.go
@@ -71,6 +71,9 @@ var (
 		vaa.ChainIDKujira,
 		vaa.ChainIDNeutron,
 		vaa.ChainIDCelestia,
+		vaa.ChainIDStargaze,
+		vaa.ChainIDSeda,
+		vaa.ChainIDDymension,
 	}
 
 	// features is the feature string to be published in the gossip heartbeat messages. It will include all chains that are actually enabled on IBC.