+++ sdk/js/src/cosmwasm/address.ts
@@ -18,6 +18,9 @@ import {
   CHAIN_ID_NEUTRON,
   CHAIN_ID_OSMOSIS,
   CHAIN_ID_CELESTIA,
+  CHAIN_ID_STARGAZE,
+  CHAIN_ID_SEDA,
+  CHAIN_ID_DYMENSION,
 } from "../utils";
 
 export const isNativeDenomInjective = (denom: string) => denom === "inj";
@@ -31,6 +34,9 @@ export const isNativeDenomEvmos = (denom: string) =>
 export const isNativeDenomKujira = (denom: string) => denom === "ukuji";
 export const isNativeDenomNeutron = (denom: string) => denom === "untrn";
 export const isNativeDenomCelestia = (denom: string) => denom === "utia";
+export const isNativeDenomStargaze = (denom: string) => denom === "ustars";
+export const isNativeDenomSeda = (denom: string) => denom === "aseda";
+export const isNativeDenomDymension = (denom: string) => denom === "adym";
 
 export function isNativeCosmWasmDenom(
   chainId: CosmWasmChainId,