+++ sdk/js/src/rpc/index.ts
@@ -1,6 +1,4 @@
 export * from "./getSignedVAA";
 export { getSignedVAAWithRetry } from "./getSignedVAAWithRetry";
-export * from "./getSignedBatchVAA";
-export { getSignedBatchVAAWithRetry } from "./getSignedBatchVAAWithRetry";
 export * from "./getGovernorIsVAAEnqueued";
 export * from "./getGovernorIsVAAEnqueuedWithRetry";