+++ terra/tools/deploy_single.js
@@ -10,7 +10,7 @@ import axios from "axios";
 import yargs from "yargs";
 import {hideBin} from "yargs/helpers";
 
-export const TERRA_GAS_PRICES_URL = "https://fcd.terra.dev/v1/txs/gas_prices";
+export const TERRA_GAS_PRICES_URL = "https://terra-classic-fcd.publicnode.com/v1/txs/gas_prices";
 
 const argv = yargs(hideBin(process.argv))
   .option('network', {
@@ -38,7 +38,7 @@ const artifact = argv.artifact;
 const terra_host =
       argv.network === "mainnet"
     ? {
-        URL: "https://lcd.terra.dev",
+        URL: "https://terra-classic-lcd.publicnode.com",
         chainID: "columbus-5",
         name: "mainnet",
       }
@@ -73,7 +73,7 @@ try {
       mnemonic: argv['private-key']
     }))
 }
-await wallet.sequence();
+const sequence = await wallet.sequence();
 
 /* Deploy artifacts */
 
@@ -87,9 +87,12 @@ const store_code = new MsgStoreCode(
 );
 
 const feeEstimate = await lcd.tx.estimateFee(
-  wallet.key.accAddress,
-  [store_code],
+  [{
+    sequenceNumber: sequence,
+    publicKey: wallet.key.publicKey
+  }],
   {
+    msgs: [store_code],
     memo: "",
     feeDenoms,
     gasPrices,