+++ sdk/js-query/CHANGELOG.md
@@ -1,6 +1,18 @@
+## 0.0.9
+
+Fix SolanaAccount mock for minContextSlot
+
+## 0.0.8
+
+Add signaturesToEvmStruct utility
+
+Fix SolanaAccount mock for dataSliceOffset == 0
+
 ## 0.0.7
 
 Add EthCallByTimestamp mock support
+
+Add SolanaAccount support
 
 ## 0.0.6