+++ sdk/js/CHANGELOG.md
@@ -1,5 +1,17 @@
 # Changelog
 
+## 0.10.11
+
+### Removed
+
+getSignedBatchVAA, getSignedBatchVAAWithRetry
+
+## 0.10.10
+
+### Changes
+
+Added CCTP Sepolia chains support
+
 ## 0.10.9
 
 ### Changes